// src/util/types.ts
export interface TopicAutomodConfig {
    characters: string[]
    length: number
    words: string[]
    default: {
        characters: string[]
        length: number
        words: string[]
    }
}

export interface SmesharikiAiConfig {
    topicmod: TopicAutomodConfig
    acceptTopics: boolean
    acceptMashups: boolean
}